excel处理统一使用EXCEL类（包括csv和database处理方式）

+++ Drill/DrillDlg.cpp
@@ -1187,7 +1187,7 @@ void CDrillDlg::GetCurNum()
                 m_nBOSeqNO = MAX(i, 1);
             if (ptTorq->bbreakout())
             {
-                m_nMaxOutWellNO = MAX(ptTorq->dwoutwellno(), m_nMaxOutWellNO);
+                m_nMaxOutWellNO = MAX((UINT)ptTorq->dwoutwellno(), m_nMaxOutWellNO);
             }
         }
     }
